feat: 계산결과를 반환하는 타입을 long 타입으로 변경
- int 타입보다 더 많은 범위의 입력을 받아서 계산할 수 있다.

+++ src/main/java/calculator/controller/CalculateController.java
@@ -15,15 +15,15 @@ public class CalculateController {
     }
 
     public void run() {
-        int result = calculateSum(inputView.inputStrings());
+        long result = calculateSum(inputView.inputStrings());
         displayResult(result);
     }
 
-    private int calculateSum(String input) {
+    private long calculateSum(String input) {
         return StringAddCalculator.splitAndSum(input);
     }
 
-    private void displayResult(int result) {
+    private void displayResult(long result) {
         outputView.printResult(result);
     }
 }
